backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.95.0
ffmpeg-python==0.2.0
filelock==3.10.0
Flask==2.2.3
Flask-Cors==3.0.10
frozenlist==1.3.3
future==0.18.3
gTTS==2.3.1
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
llvmlite==0.39.1
MarkupSafe==2.1.2
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.0
numba==0.56.4
numpy==1.23.5
openai==0.27.2
openai-whisper==20230314
packaging==23.0
pydantic==1.10.6
python-certifi-win32==1.6.1
python-multipart==0.0.6
regex==2022.10.31
requests==2.28.2
setuptools-scm==7.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
starlette==0.26.1
sympy==1.11.1
tiktoken==0.3.1
tomli==2.0.1
torch==2.0.0
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
Werkzeug==2.2.3
wrapt==1.15.0
yarl==1.8.2
python-dotenv

